+++ packages/coding-agent/test/suite/regressions/5943-session-start-notify.test.ts
@@ -18,6 +18,7 @@ function createUiContext(
 		onTerminalInput: () => () => {},
 		getTranscriptSelection: () => undefined,
 		onTranscriptSelection: () => () => {},
+		setTranscriptAnnotations: () => {},
 		setStatus: () => {},
 		setWorkingMessage: () => {},
 		setWorkingVisible: () => {},